refactoring, create Fom component
